Allow extending builder pattern
